Hook examples.
