--- v3/tasks/20241212-v3-hook-examples/readme.md
@@ -1,10 +1,0 @@
-# 2024-12-12 - V3 Hook examples
-
-Test deployment for fee taking hook example. Uses Vault's 11th release.
-Code [link](https://github.com/balancer/balancer-v3-monorepo/commit/73708b75898a62dac0535f38d1bf471ac0e538c6).
-
-## Useful Files
-
-- [Sepolia testnet addresses](./output/sepolia.json)
-- [`FeeTakingHookExample` artifact](./artifact/FeeTakingHookExample.json)
-
