+++ v3/tasks/20241213-v3-hook-examples/readme.md
@@ -1,0 +1,14 @@
+# 2024-12-13 - V3 Hook examples
+
+Hook examples deployment, showcasing how to create and plug in simple hooks to pools.
+Contracts deployed in this task are mostly for educational and testing purposes; they are not meant to be used in a production environment.
+
+## Useful Files
+
+- Code [link](https://github.com/balancer/balancer-v3-monorepo/commit/2f088c6b8f66ad55885d257c1e3debe2a6e21e97).
+- [`DirectionalFeeHookExample` artifact](./artifact/DirectionalFeeHookExample.json)
+- [`ExitFeeHookExample` artifact](./artifact/ExitFeeHookExample.json)
+- [`FeeTakingHookExample` artifact](./artifact/FeeTakingHookExample.json)
+- [`LotteryHookExample` artifact](./artifact/LotteryHookExample.json)
+- [`NftLiquidityPositionExample` artifact](./artifact/NftLiquidityPositionExample.json)
+- [`VeBALFeeDiscountHookExample` artifact](./artifact/VeBALFeeDiscountHookExample.json)
